optimized current solution + starting to implement unity hdrp capsule shadows
optimized current solution + starting to implement the unity hdrp capsule shadows implementation. doing it so I have two different versions that are implemented. One that originally was made for HDRP but never included in official releases (at the time of writing). The goal is with the 2 versions implemented to make a version that combines the best of both versions to make the best possible version.